fix(#137): 매직넘버 공식 수정 및 시뮬레이션 오버플로우 방지
- 표준 야구 매직넘버 공식으로 수정
  - 기존: leader.remainingGames + 1 - (leader.wins - stats.wins)
  - 변경: stats.remainingGames - leader.wins + stats.wins + 1
  - 표준 공식은 상대팀(2위) 기준 남은경기를 사용해야 함
- Int → Long 변경으로 완전 탐색 시뮬레이션 오버플로우 방지
  - totalScenarios, qualifyingScenarios 필드 Long으로 변경
  - MONTE_CARLO_ITERATIONS 상수 Long으로 변경
  - 완전 탐색 루프 변수를 Long으로 변경
- 테스트 기대값 업데이트 (Long 리터럴 및 공식 변경 반영)

+++ nextup-api/src/main/kotlin/com/nextup/api/dto/standings/SimulationApiDto.kt
@@ -103,8 +103,8 @@ data class SimulationResultResponse(
  * 플레이오프 시나리오 결과 응답
  */
 data class PlayoffScenarioResponse(
-    val totalScenarios: Int,
-    val qualifyingScenarios: Int,
+    val totalScenarios: Long,
+    val qualifyingScenarios: Long,
     val probability: Double,
     val magicNumber: Int?,
 ) {
